requirements.txt
speedtest-cli>=2.1.3
matplotlib>=3.5.0
numpy>=1.21.0
requests>=2.25.0
